+++ SomnusLoadBalancer/src/main/java/com/somnus/server/somnuslb/exception/SomnusException.java
@@ -1,0 +1,8 @@
+package com.somnus.server.somnuslb.exception;
+
+public class SomnusException extends RuntimeException {
+
+    public SomnusException(String message) {
+        super(message);
+    }
+}