+++ SomnusLoadBalancer/src/main/java/com/somnus/server/somnuslb/mirrorHealth/domain/MirrorState.java
@@ -1,0 +1,8 @@
+package com.somnus.server.somnuslb.mirrorHealth.domain;
+
+public enum MirrorState {
+    NEW_MIRROR,
+    NORMAL_OPERATION,
+    MIRROR_DOWN,
+    MIRROR_RECOVERY
+}