+++ SomnusLoadBalancer/src/main/java/com/somnus/server/somnuslb/mirrors/MirrorInfoService.java
@@ -1,0 +1,14 @@
+package com.somnus.server.somnuslb.mirrors;
+
+import com.somnus.server.somnuslb.observer.Observer;
+import org.springframework.stereotype.Service;
+
+@Service
+public class MirrorInfoService extends MirrorObservable {
+
+    public MirrorInfoService() { }
+
+    public void addMirror() {
+        // TODO
+    }
+}