+++ SomnusLoadBalancer/src/main/java/com/somnus/server/somnuslb/observer/Observable.java
@@ -1,0 +1,8 @@
+package com.somnus.server.somnuslb.observer;
+
+public interface Observable {
+    public void addObserver(Observer o);
+    public void removeObserver(Observer o);
+    public void notifyAddition(Object o);
+    public void notifyRemoval(Object o);
+}