+++ SomnusLoadBalancer/src/main/java/com/somnus/server/somnuslb/observer/Observer.java
@@ -1,0 +1,6 @@
+package com.somnus.server.somnuslb.observer;
+
+public interface Observer {
+    public void add(Object o);
+    public void remove(Object o);
+}